Backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# AI Integration
google-generativeai==0.3.1

# Configuration
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# HTTP Client
httpx==0.25.2

# Email validation
email-validator==2.1.0

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Code quality (optional)
black==23.11.0
flake8==6.1.0
mypy==1.7.1
